fix: svelte & vue
