+++ examples/app-frameworks/fastapi/frontend/package.json
@@ -13,6 +13,7 @@
     "@openuidev/react-lang": "0.2.15",
     "@openuidev/react-ui": "0.13.10",
     "@openuidev/thesys": "0.3.2",
+    "ai": "^7.0.85",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "zod": "^4.0.0",
