+++ examples/app-frameworks/svelte/package.json
@@ -3,9 +3,10 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "vite dev",
-    "build": "vite build",
-    "verify": "vite build",
+    "prepare": "svelte-kit sync || echo ''",
+    "dev": "svelte-kit sync && vite dev",
+    "build": "svelte-kit sync && vite build",
+    "verify": "svelte-kit sync && vite build",
     "preview": "vite preview"
   },
   "dependencies": {
